hoist useColorOverrideReducer
