+++ packages/react-components/theme-designer/src/components/ColorTokens/OverridableTokenBrandColors.ts
@@ -1,5 +1,5 @@
 import type { Brands, BrandVariants, Theme } from '@fluentui/react-theme';
-import { ColorOverrideBrands } from './useColorOverrideReducer';
+import { ColorOverrideBrands } from './ColorTokens';
 
 export const brandRamp: Brands[] = [10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 110, 120, 130, 140, 150, 160];
