+++ packages/react-components/theme-designer/src/useThemeDesignerReducer.ts
@@ -44,44 +44,6 @@ export type AppState = {
   overrides: Partial<Theme>;
 };
 
-const overrideReducer = (state: OverrideState, action: { type: string; overrides?: Partial<Theme> }) => {
-  if (!action.overrides) {
-    switch (action.type) {
-      case 'Teams Light':
-        return { ...state, teamsLight: {} };
-      case 'Teams Dark':
-        return { ...state, teamsDark: {} };
-      case 'Web Light':
-        return { ...state, webLight: {} };
-      case 'Web Dark':
-        return { ...state, webDark: {} };
-      case 'Custom Light':
-        return { ...state, customLight: {} };
-      case 'Custom Dark':
-        return { ...state, customDark: {} };
-      default:
-        return state;
-    }
-  } else {
-    switch (action.type) {
-      case 'Teams Light':
-        return { ...state, teamsLight: { ...state.teamsLight, ...action.overrides } };
-      case 'Teams Dark':
-        return { ...state, teamsDark: { ...state.teamsDark, ...action.overrides } };
-      case 'Web Light':
-        return { ...state, webLight: { ...state.webLight, ...action.overrides } };
-      case 'Web Dark':
-        return { ...state, webDark: { ...state.webDark, ...action.overrides } };
-      case 'Custom Light':
-        return { ...state, customLight: { ...state.customLight, ...action.overrides } };
-      case 'Custom Dark':
-        return { ...state, customDark: { ...state.customDark, ...action.overrides } };
-      default:
-        return state;
-    }
-  }
-};
-
 export const useThemeDesignerReducer = () => {
   const createCustomTheme = ({
     darkCp,
@@ -101,16 +63,64 @@ export const useThemeDesignerReducer = () => {
     };
   };
 
-  const [overrideState, dispatchOverrideState] = React.useReducer(overrideReducer, {
+  const initialOverrideState = {
     teamsLight: {},
     teamsDark: {},
     webLight: {},
     webDark: {},
     customLight: {},
     customDark: {},
-  });
+  };
 
-  const stateReducer = (state: AppState, action: DispatchTheme) => {
+  const overrideReducer = (state: OverrideState, action: { type: string; overrides?: Partial<Theme> }) => {
+    if (!action.overrides) {
+      switch (action.type) {
+        case 'Teams Light':
+          return { ...state, teamsLight: {} };
+        case 'Teams Dark':
+          return { ...state, teamsDark: {} };
+        case 'Web Light':
+          return { ...state, webLight: {} };
+        case 'Web Dark':
+          return { ...state, webDark: {} };
+        case 'Custom Light':
+          return { ...state, customLight: {} };
+        case 'Custom Dark':
+          return { ...state, customDark: {} };
+        default:
+          return state;
+      }
+    } else {
+      switch (action.type) {
+        case 'Teams Light':
+          return { ...state, teamsLight: { ...state.teamsLight, ...action.overrides } };
+        case 'Teams Dark':
+          return { ...state, teamsDark: { ...state.teamsDark, ...action.overrides } };
+        case 'Web Light':
+          return { ...state, webLight: { ...state.webLight, ...action.overrides } };
+        case 'Web Dark':
+          return { ...state, webDark: { ...state.webDark, ...action.overrides } };
+        case 'Custom Light':
+          return { ...state, customLight: { ...state.customLight, ...action.overrides } };
+        case 'Custom Dark':
+          return { ...state, customDark: { ...state.customDark, ...action.overrides } };
+        default:
+          return state;
+      }
+    }
+  };
+
+  const [overrideState, dispatchOverrideState] = React.useReducer(overrideReducer, initialOverrideState);
+
+  const initialAppState = {
+    themeLabel: 'Teams Light',
+    brand: brandTeams,
+    theme: teamsLightTheme,
+    isDark: false,
+    overrides: {},
+  };
+
+  const appStateReducer = (state: AppState, action: DispatchTheme) => {
     if (action.type === 'Override') {
       if (action.overrides) {
         if (state.themeLabel === 'Custom') {
@@ -179,11 +189,5 @@ export const useThemeDesignerReducer = () => {
     }
   };
 
-  return React.useReducer(stateReducer, {
-    themeLabel: 'Teams Light',
-    brand: brandTeams,
-    theme: teamsLightTheme,
-    isDark: false,
-    overrides: {},
-  });
+  return React.useReducer(appStateReducer, initialAppState);
 };
